requirements-socketio.txt
# Socket.IO dependencies
python-socketio>=5.13.0
python-engineio>=4.12.0
aiohttp>=3.8.5
